create project service and fixes

+++ src/main/java/com/crowdfunding/dp/backend/dao/UserRepository.java
@@ -4,7 +4,7 @@ import com.crowdfunding.dp.backend.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.Repository;
 
-public interface UserRepository extends JpaRepository<User, Integer>, Repository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long>, Repository<User, Long> {
 
     User findUserByLoginEquals(String login);
 }
